+++ packages/admin/react-router.config.ts
@@ -1,9 +1,5 @@
 import type { Config } from "@react-router/dev/config";
 
 export default {
-  async prerender() {
-    return ["/", "/admin"];
-  },
-  // Server-side render by default, to enable SPA mode set this to `false`
   ssr: false,
 } satisfies Config;