+++ packages/server/CHANGELOG.md
@@ -1,11 +1,29 @@
 # @isardsat/editorial-server
 
+## 6.0.6
+
+### Patch Changes
+
+- 24fb391: Add url to zod types
+- Updated dependencies [24fb391]
+  - @isardsat/editorial-common@6.0.6
+  - @isardsat/editorial-admin@6.0.6
+
+## 6.0.5
+
+### Patch Changes
+
+- bcb0c8a: Fix admin URLs
+- Updated dependencies [bcb0c8a]
+  - @isardsat/editorial-common@6.0.5
+  - @isardsat/editorial-admin@6.0.5
+
 ## 6.0.4
 
 ### Patch Changes
 
-- 704dbf7: Build re-release
-- Updated dependencies [704dbf7]
+- 4936462: Fix broken release
+- Updated dependencies [4936462]
   - @isardsat/editorial-common@6.0.4
   - @isardsat/editorial-admin@6.0.4
 
@@ -13,8 +31,8 @@
 
 ### Patch Changes
 
-- d161b48: Initial i18n integration
-- Updated dependencies [d161b48]
+- 0c196b0: Fix build errors
+- Updated dependencies [0c196b0]
   - @isardsat/editorial-common@6.0.3
   - @isardsat/editorial-admin@6.0.3